fix: split fork api health server from core
